start refactoring the whole landing page
