--- client/components/home/sections/index.ts
@@ -1,8 +1,0 @@
-import Api from "./api";
-import Banner from "./banner";
-import Customisation from "./customisation";
-import Ending from "./ending";
-import Moderation from "./moderation";
-import PlainHTML from "./plainHTML";
-
-export { Api, Banner, Customisation, Ending, Moderation, PlainHTML };
